Declare in one place which models this repository carries
A corpus was owned by whichever module first needed it, so which side of the
compiler a module sat on decided what could be asked of its models. Everything
that wanted a set of them from the other side wrote the membership out again,
and those writings answer separately with nothing failing when they differ: one
of them had lost a file of a corpus, and another had lost a corpus.

The models now live in souther-test-support, which every module already reaches
whichever side it is on, and `corpora.txt` names each of them with what it was
written for. That, with each corpus's `sources.txt`, is the whole of what says
which files a corpus is made of; every reader derives its set from the two and
names no member of its own. What a corpus was written for is a line of the
manifest rather than the directory it sits in, so placement decides nothing
again.

What this compiler answers about a conformance corpus stays beside the compiler.
Those are answers about a model rather than part of it, and the one caller that
writes them back owns where they are written.

The measurement jar carries the models without depending on the module that owns
them: it copies the resources into its own artifact, the way the compiler
carries the specification, so a measurement replayed from an old jar still finds
the corpus of its revision.

Two checks see something they could not see before, because reaching a corpus by
walking repository paths is invisible to the check that reads what sweeps a
population. One of them really was sweeping under a tag on a single method, which
surefire reads and that check cannot: the sweep is its own class now, and the
structural questions beside it stay in the run somebody waits on. The other
draws one model from the corpora rather than quantifying over them, which is what
the list of checks asked anyway is for — and its note now says that is the
criterion, rather than describing the two entries it happened to hold.
