--- souther-compiler/src/test/resources/souther/compiler/conformance/corpora.txt
@@ -1,3 +1,0 @@
-booking
-catalog
-staffing
