+++ souther-test-support/src/main/resources/souther/corpus/catalog/shipping.examples.sou
