+++ souther-test-support/src/main/resources/souther/corpus/corpora.txt
@@ -1,0 +1,29 @@
+# Every model this repository carries, and what each was written for.
+#
+# One line per corpus: the directory it is in, then what it was written for. This is the whole of
+# what says a corpus exists, and each corpus's sources.txt is the whole of what says which files it
+# is made of and in what order. Anything wanting a set of models reads these two and names no member
+# of its own. A second list of the same members drifts from this one without anything going red:
+# this repository had one that dropped a file and one that dropped a corpus, and neither was.
+#
+# What it was written for is here rather than in the directory a corpus sits in, because that is
+# what this is about. A corpus grouped by its use would say what it is for twice, and the placement
+# would start deciding what may be asked of it again.
+#
+#   conformance   Written against what the language declares. What this compiler answers about it
+#                 is checked in beside the compiler, and it grows whenever a rule is tightened.
+#   measurement   Written as somebody would write a model of this size, and a number is compared
+#                 against it. It must not move: a corpus that grew between two revisions makes the
+#                 second one look slower.
+#   running       There to be run rather than to be compiled. Timing its compile would say nothing;
+#                 what it is for is that the classes it generates are loaded and called.
+#
+# The order is the order they are handed out in. The measurement corpora are largest first, which is
+# the order a run reports them in.
+
+booking conformance
+catalog conformance
+staffing conformance
+crm measurement
+issuetracker measurement
+runtime running
