+++ souther-test-support/src/main/resources/souther/corpus/crm/pipeline.sou
